# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Manager API",
        "description": "API Documentation for the Project Management System. <br> **Base URL**: /api",
        "contact": {
            "email": "admin@company.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List all projects",
                "description": "Executives see all projects. Managers and Associates see only projects they are assigned to (via Team or Advisory).",
                "operationId": "getProjects",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of projects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "description": "Creates a new project record. Restricted to Executives.",
                "operationId": "createProject",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Q4 Marketing Refresh"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Overhaul of the main landing pages."
                                    },
                                    "status": {
                                        "type": "string",
                                        "default": "active",
                                        "enum": [
                                            "active",
                                            "hold",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden. Executives only."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/projects/{project}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get project details",
                "description": "View details, including assigned teams and advisors. Users must have access rights.",
                "operationId": "getProjectById",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "Project ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update project details",
                "description": "Update metadata or status. Executives only.",
                "operationId": "updateProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "hold",
                                            "completed",
                                            "archived"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Archive/Delete project",
                "description": "Soft deletes the project. Executives only.",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project archived",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Project archived"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProjectResource": {
                "title": "Project Resource",
                "description": "Details of a project including metrics, assigned teams, and advisors.",
                "properties": {
                    "id": {
                        "description": "Unique Project ID",
                        "type": "integer",
                        "example": 101
                    },
                    "name": {
                        "description": "Project Name",
                        "type": "string",
                        "example": "Q3 Marketing Campaign"
                    },
                    "description": {
                        "type": "string",
                        "example": "Revamping the core legacy system.",
                        "nullable": true
                    },
                    "status": {
                        "description": "Current status of the project",
                        "type": "string",
                        "enum": [
                            "active",
                            "completed",
                            "archived",
                            "hold"
                        ],
                        "example": "active"
                    },
                    "metrics": {
                        "description": "Aggregated review statistics",
                        "properties": {
                            "average_rating": {
                                "type": "number",
                                "format": "float",
                                "example": 4.5
                            },
                            "review_count": {
                                "type": "integer",
                                "example": 12
                            }
                        },
                        "type": "object"
                    },
                    "created_by": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ],
                        "nullable": true,
                        "description": "The Executive who created the project (nullable if user deleted)"
                    },
                    "teams": {
                        "description": "List of teams assigned to this project",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TeamResource"
                        }
                    },
                    "advisors": {
                        "description": "List of internal advisors assigned to this project",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-31T23:59:59Z",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "TeamResource": {
                "title": "Team Resource",
                "description": "Represents a team within the organization.",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the team",
                        "type": "integer",
                        "example": 10
                    },
                    "name": {
                        "description": "Name of the team",
                        "type": "string",
                        "example": "Frontend Alpha"
                    },
                    "member_count": {
                        "description": "Total number of members in the team (Conditional)",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "members": {
                        "description": "List of team members (Conditional: only when roster is loaded)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        },
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Timestamp when the team was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "User Resource",
                "description": "Employee details including global roles and context-specific pivot data.",
                "properties": {
                    "id": {
                        "description": "Unique User ID",
                        "type": "integer",
                        "example": 101
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Alice"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Smith"
                    },
                    "full_name": {
                        "description": "Computed full name attribute",
                        "type": "string",
                        "readOnly": true,
                        "example": "Alice Smith"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "alice@company.com"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "roles": {
                        "description": "List of assigned global roles (Executive, Manager, Associate)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Associate"
                        }
                    },
                    "joined_team_at": {
                        "description": "Timestamp when user joined the team. Only present in Team Roster views.",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "assigned_advisor_at": {
                        "description": "Timestamp when user was assigned as an advisor. Only present in Project Advisor contexts.",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}